+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - Fix unexpected navigation in preview panel.
 - Add support of project root in run configurations.
-
+- Add support of tinymist status indicator
 ## [0.1.0]
 
 All major feature is implemented.